feat(kp-cloud): 补充订单所有信息

+++ kp-cloud/kp-system/src/main/java/org/dromara/kp/system/domain/KpDiscountActivity.java
@@ -56,12 +56,12 @@ public class KpDiscountActivity extends TenantEntity {
     /**
      * 禁用
      */
-    private Long disableFlag;
+    private Integer disableFlag;
 
     /**
      * 活动类型
      */
-    private Long activityType;
+    private Integer activityType;
 
     /**
      * 备注
